requirements.txt
appdirs==1.4.4
attrs==23.2.0
backoff==2.2.1
boilerpy3==1.0.7
canals==0.3.2
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docopt==0.6.2
Events==0.5
exceptiongroup==1.2.0
faiss-cpu==1.7.2
farm-haystack==1.19.0
filelock==3.13.1
fsspec==2024.2.0
greenlet==3.0.3
huggingface-hub==0.21.4
idna==3.6
inflect==7.0.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pip==24.0
platformdirs==4.2.0
posthog==3.5.0
prompthub-py==4.0.0
pydantic==1.10.14
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==58.1.0
six==1.16.0
SQLAlchemy==1.4.52
SQLAlchemy-Utils==0.41.1
sseclient-py==1.8.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchvision==0.17.1+cu121
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1